refactor: extract all fns to their own c file
